Gather top-level commands
