+++ micropip/commands/uninstall.py
@@ -3,8 +3,8 @@ import importlib.metadata
 import warnings
 from importlib.metadata import Distribution
 
-from ._compat import loadedPackages
-from ._utils import get_files_in_distribution, get_root
+from .._compat import loadedPackages
+from .._utils import get_files_in_distribution, get_root
 
 
 def uninstall(packages: str | list[str]) -> None:
